+++ app/src/main/java/com/example/workoutlog/interfaces/ISetDao.java
@@ -37,7 +37,7 @@ public interface ISetDao {
 
     @Transaction
     @Query("SELECT MAX(set_table.weight) as weight FROM set_table,user_routine_exercise_table,exercise_table,workout_table WHERE set_table.userRoutineExerciseRoutineId = user_routine_exercise_table.id AND user_routine_exercise_table.exerciseTypeId = :exerciseId AND user_routine_exercise_table.workoutId= :workoutId")
-    float getMaxWeightFromWorkoutWithExercise(long workoutId, long exerciseId);
+    double getMaxWeightFromWorkoutWithExercise(long workoutId, long exerciseId);
 
 
     @Transaction
